requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiohttp-retry==2.9.1
aiosignal==1.3.2
amqp==5.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
anyscale==0.26.33
appdirs==1.4.4
argcomplete==3.6.2
async-timeout==5.0.1
asyncssh==2.21.0
atpublic==6.0.1
attrs==25.3.0
billiard==4.2.1
black==25.1.0
boto3==1.38.41
botocore==1.38.41
cachetools==5.5.2
celery==5.5.3
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
configobj==5.0.9
cryptography==45.0.4
datamodel-code-generator==0.26.1
decorator==5.2.1
dictdiffer==0.9.0
diskcache==5.6.3
distro==1.9.0
dnspython==2.7.0
dpath==2.2.0
dulwich==0.23.0
dvc==3.60.1
dvc-data==3.16.10
dvc-gs==3.0.2
dvc-http==2.32.0
dvc-objects==5.1.1
dvc-render==1.0.2
dvc-studio-client==0.21.0
dvc-task==0.40.2
email_validator==2.2.0
entrypoints==0.4
exceptiongroup==1.3.0
Faker==37.4.0
filelock==3.18.0
flatten-dict==0.4.2
flufl.lock==8.2.0
frozenlist==1.7.0
fsspec==2025.5.1
funcy==2.0
gcsfs==2025.5.1
genson==1.3.0
gitdb==4.0.12
GitPython==3.1.44
google-api-core==2.25.1
google-api-python-client==2.173.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
google-cloud-certificate-manager==1.10.2
google-cloud-common==1.5.2
google-cloud-compute==1.31.0
google-cloud-core==2.4.3
google-cloud-filestore==1.13.2
google-cloud-redis==2.18.1
google-cloud-resource-manager==1.14.2
google-cloud-secret-manager==2.24.0
google-cloud-storage==3.1.1
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grandalf==0.8
grpc-google-iam-v1==0.14.2
grpcio==1.73.0
grpcio-status==1.48.2
gto==1.7.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
humanize==4.12.3
hydra-core==1.3.2
idna==3.10
ijson==3.4.0
inflect==5.6.2
isort==5.13.2
iterative-telemetry==0.0.10
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.1
jsf==0.11.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kombu==5.5.4
label-studio-sdk==1.0.18
log-symbols==0.0.14
lxml==5.4.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
multidict==6.5.0
mypy_extensions==1.1.0
networkx==3.4.2
nltk==3.9.1
numpy==2.2.6
oauth2client==4.1.3
oauthlib==3.3.1
omegaconf==2.3.0
opencv-python==4.11.0.86
orjson==3.10.18
packaging==25.0
pandas==2.3.0
pathspec==0.12.1
pillow==11.2.1
platformdirs==4.3.8
prompt_toolkit==3.0.51
propcache==0.3.2
proto-plus==1.26.1
protobuf==3.20.3
psutil==7.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydot==4.0.1
pygit2==1.18.0
Pygments==2.19.2
pygtrie==2.5.0
PyJWT==2.10.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
requests-mock==1.12.1
requests-oauthlib==2.0.0
rich==14.0.0
rpds-py==0.25.1
rsa==4.9.1
rstr==3.2.2
ruamel.yaml==0.18.14
ruamel.yaml.clib==0.2.12
s3transfer==0.13.0
scmrepo==3.3.11
semver==3.0.4
shellingham==1.5.4
shortuuid==1.0.13
shtab==1.7.2
six==1.17.0
smart-open==7.1.0
smmap==5.0.2
sniffio==1.3.1
spinners==0.0.24
sqltrie==0.11.2
tabulate==0.9.0
termcolor==3.1.0
toml==0.10.2
tomli==2.2.1
tomlkit==0.13.3
tqdm==4.67.1
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
tzlocal==5.3.1
ujson==5.10.0
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
voluptuous==0.15.2
wcwidth==0.2.13
websockets==15.0.1
wrapt==1.17.2
xmljson==0.2.1
yarl==1.20.1
zc.lockfile==3.0.post1
